Add new server cli class
This adds a new CLI class to run the server. We have also switched from
shotgun to rerun, so that changes to the filesystem are detected and
the app will automatically reload when detected. This is a better
approach than shotgun, because shotgun was reloading the entire app on
every request.

Fixes #3
Fixes #9